Updating Recharts branch to rebased version of master

+++ src/main/java/com/javapid/objects/recharts/AdultObject.java
@@ -2,7 +2,7 @@ package com.javapid.objects.recharts;
 
 public class AdultObject extends PersonAbstractClass {
 
-	public AdultObject(String month, Integer ticketCount) {
+	public AdultObject(String month, Long ticketCount) {
 		setName("Dospelý");
 		setMonth(month);
 		setValue(ticketCount);
