+++ src/main/java/com/javapid/objects/recharts/JuniorObject.java
@@ -2,7 +2,7 @@ package com.javapid.objects.recharts;
 
 public class JuniorObject extends PersonAbstractClass {
 
-	public JuniorObject(String month, Integer ticketCount){
+	public JuniorObject(String month, Long ticketCount){
 		setName("Junior");
 		setMonth(month);
 		setValue(ticketCount);
