+++ src/main/java/com/javapid/objects/recharts/PortableObject.java
@@ -2,7 +2,7 @@ package com.javapid.objects.recharts;
 
 public class PortableObject extends PersonAbstractClass {
 
-    public PortableObject(String month, Integer ticketCount){
+    public PortableObject(String month, Long ticketCount){
         setName("Portable");
         setMonth(month);
         setValue(ticketCount);
