+++ src/main/java/com/javapid/objects/recharts/SeniorObject.java
@@ -2,7 +2,7 @@ package com.javapid.objects.recharts;
 
 public class SeniorObject extends PersonAbstractClass {
 
-    public SeniorObject(String month, Integer ticketCount){
+    public SeniorObject(String month, Long ticketCount){
         setName("Senior");
         setMonth(month);
         setValue(ticketCount);
