+++ src/main/java/com/javapid/objects/recharts/StudentObject.java
@@ -2,7 +2,7 @@ package com.javapid.objects.recharts;
 
 public class StudentObject extends PersonAbstractClass {
 
-    public StudentObject(String month, Integer ticketCount){
+    public StudentObject(String month, Long ticketCount){
         setName("Student");
         setMonth(month);
         setValue(ticketCount);
